Create order-of-compilation.md
Hi, I am maybefbi from the HypeIRC forums. This is the order of compilation of in the CJDNS project.

I created it using a btrace (https://github.com/rprichard/sourceweb/blob/master/btrace/sw-btrace) which loads a shared object into a process which logs to the console when new processes are spawned from it. I parsed the JSON output it generated and looked for spawns of gcc with -c, -S or -E flags.